+++ scifit/tests/test_solvers_specials.py
@@ -14,10 +14,10 @@ from scifit.tests.test_solvers_linear import GenericLinearRegression
 
 class DebyeRegression(GenericLinearRegression):
     factory = specials.DebyeFitSolver
-    parameters = np.array([428.])
+    parameters = np.array([428.0])
     sigma = None
-    xmin = 25.
-    xmax = 550.
+    xmin = 25.0
+    xmax = 550.0
 
 
 class DebyeRegressionNoiseL0(DebyeRegression, TestCase):
@@ -25,5 +25,28 @@ class DebyeRegressionNoiseL0(DebyeRegression, TestCase):
 
 
 class DebyeRegressionNoiseL1(DebyeRegression, TestCase):
-    sigma = 1e-2
+    sigma = 5e-2
 
+
+class CrankDiffusionRegression(GenericLinearRegression):
+    factory = specials.CrankDiffusionFitSolver
+    parameters = np.array([3.9, 2e-11])
+    configuration = {"p0": np.array([5.0, 1e-10])}
+    sigma = None
+    xmin = 1e0
+    xmax = 1e6
+    resolution = 20
+    mode = "log"
+    log_x = True
+    log_y = True
+    log_loss = True
+    loss_resolution = 20
+
+
+class CrankDiffusionRegressionNoiseL0(CrankDiffusionRegression, TestCase):
+    sigma = 1e-6
+
+
+# class CrankDiffusionRegressionNoiseL1(CrankDiffusionRegression, TestCase):
+#     sigma = 5e-2
+#